Revert move of UserAttributes to keep MSAL free of changes
