+++ common4j/src/main/com/microsoft/identity/common/java/nativeauth/controllers/results/SignUpCommandResult.kt
@@ -22,7 +22,7 @@
 // THE SOFTWARE.
 package com.microsoft.identity.common.java.nativeauth.controllers.results
 
-import com.microsoft.identity.common.java.nativeauth.providers.responses.signup.UserAttributeApiResult
+import com.microsoft.identity.common.java.nativeauth.providers.responses.UserAttributeApiResult
 
 sealed interface SignUpSubmitCodeCommandResult: INativeAuthCommandResult
 sealed interface SignUpSubmitUserAttributesCommandResult: INativeAuthCommandResult
