+++ common4j/src/main/com/microsoft/identity/common/java/nativeauth/providers/responses/UserAttributeOptionsApiResult.kt
@@ -20,7 +20,7 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.common.java.nativeauth.providers.responses.signup
+package com.microsoft.identity.common.java.nativeauth.providers.responses
 
 import com.google.gson.annotations.SerializedName
 import com.microsoft.identity.common.java.nativeauth.util.ILoggable
